+++ egui-winit/src/epi.rs
@@ -193,11 +193,13 @@ pub struct EpiIntegration {
     pub app: Box<dyn epi::App>,
     /// When set, it is time to quit
     quit: bool,
+    can_drag_window: bool,
 }
 
 impl EpiIntegration {
     pub fn new(
         integration_name: &'static str,
+        max_texture_side: usize,
         window: &winit::window::Window,
         repaint_signal: std::sync::Arc<dyn epi::backend::RepaintSignal>,
         persistence: crate::epi::Persistence,
@@ -223,9 +225,10 @@ impl EpiIntegration {
             frame,
             persistence,
             egui_ctx,
-            egui_winit: crate::State::new(window),
+            egui_winit: crate::State::new(max_texture_side, window),
             app,
             quit: false,
+            can_drag_window: false,
         };
 
         slf.setup(window);
@@ -263,11 +266,17 @@ impl EpiIntegration {
     }
 
     pub fn on_event(&mut self, event: &winit::event::WindowEvent<'_>) {
-        use winit::event::WindowEvent;
-        if *event == WindowEvent::CloseRequested {
-            self.quit = self.app.on_exit_event();
-        } else if *event == WindowEvent::Destroyed {
-            self.quit = true;
+        use winit::event::{ElementState, MouseButton, WindowEvent};
+
+        match event {
+            WindowEvent::CloseRequested => self.quit = self.app.on_exit_event(),
+            WindowEvent::Destroyed => self.quit = true,
+            WindowEvent::MouseInput {
+                button: MouseButton::Left,
+                state: ElementState::Pressed,
+                ..
+            } => self.can_drag_window = true,
+            _ => {}
         }
 
         self.egui_winit.on_event(&self.egui_ctx, event);
@@ -290,7 +299,9 @@ impl EpiIntegration {
             .egui_winit
             .handle_output(window, &self.egui_ctx, egui_output);
 
-        let app_output = self.frame.take_app_output();
+        let mut app_output = self.frame.take_app_output();
+        app_output.drag_window &= self.can_drag_window; // Necessary on Windows; see https://github.com/emilk/egui/pull/1108
+        self.can_drag_window = false;
 
         if app_output.quit {
             self.quit = self.app.on_exit_event();