+++ egui/src/context.rs
@@ -2,9 +2,9 @@
 
 use crate::{
     animation_manager::AnimationManager, data::output::Output, frame_state::FrameState,
-    input_state::*, layers::GraphicLayers, TextureHandle, *,
+    input_state::*, layers::GraphicLayers, memory::Options, TextureHandle, *,
 };
-use epaint::{mutex::*, stats::*, text::Fonts, *};
+use epaint::{mutex::*, stats::*, text::Fonts, TessellationOptions, *};
 
 // ----------------------------------------------------------------------------
 
@@ -33,7 +33,6 @@ struct ContextImpl {
     fonts: Option<Fonts>,
     memory: Memory,
     animation_manager: AnimationManager,
-    latest_font_image_version: Option<u64>,
     tex_manager: WrappedTextureManager,
 
     input: InputState,
@@ -63,7 +62,7 @@ impl ContextImpl {
         self.input = input.begin_frame(new_raw_input);
         self.frame_state.begin_frame(&self.input);
 
-        self.update_fonts_mut(self.input.pixels_per_point());
+        self.update_fonts_mut();
 
         // Ensure we register the background area so panels and background ui can catch clicks:
         let screen_rect = self.input.screen_rect();
@@ -78,26 +77,28 @@ impl ContextImpl {
     }
 
     /// Load fonts unless already loaded.
-    fn update_fonts_mut(&mut self, pixels_per_point: f32) {
-        let new_font_definitions = self.memory.new_font_definitions.take();
+    fn update_fonts_mut(&mut self) {
+        let pixels_per_point = self.input.pixels_per_point();
+        let max_texture_side = self.input.raw.max_texture_side;
 
-        let pixels_per_point_changed = match &self.fonts {
-            None => true,
-            Some(current_fonts) => {
-                (current_fonts.pixels_per_point() - pixels_per_point).abs() > 1e-3
+        if let Some(font_definitions) = self.memory.new_font_definitions.take() {
+            let fonts = Fonts::new(pixels_per_point, max_texture_side, font_definitions);
+            self.fonts = Some(fonts);
+        }
+
+        let fonts = self.fonts.get_or_insert_with(|| {
+            let font_definitions = FontDefinitions::default();
+            Fonts::new(pixels_per_point, max_texture_side, font_definitions)
+        });
+
+        fonts.begin_frame(pixels_per_point, max_texture_side);
+
+        if self.memory.options.preload_font_glyphs {
+            // Preload the most common characters for the most common fonts.
+            // This is not very important to do, but may a few GPU operations.
+            for font_id in self.memory.options.style.text_styles.values() {
+                fonts.lock().fonts.font(font_id).preload_common_characters();
             }
-        };
-
-        if self.fonts.is_none() || new_font_definitions.is_some() || pixels_per_point_changed {
-            self.fonts = Some(Fonts::new(
-                pixels_per_point,
-                new_font_definitions.unwrap_or_else(|| {
-                    self.fonts
-                        .as_ref()
-                        .map(|font| font.definitions().clone())
-                        .unwrap_or_default()
-                }),
-            ));
         }
     }
 }
@@ -443,20 +444,31 @@ impl Context {
 /// ## Borrows parts of [`Context`]
 impl Context {
     /// Stores all the egui state.
+    ///
     /// If you want to store/restore egui, serialize this.
+    #[inline]
     pub fn memory(&self) -> RwLockWriteGuard<'_, Memory> {
         RwLockWriteGuard::map(self.write(), |c| &mut c.memory)
     }
 
+    /// Stores superficial widget state.
+    #[inline]
+    pub fn data(&self) -> RwLockWriteGuard<'_, crate::util::IdTypeMap> {
+        RwLockWriteGuard::map(self.write(), |c| &mut c.memory.data)
+    }
+
+    #[inline]
     pub(crate) fn graphics(&self) -> RwLockWriteGuard<'_, GraphicLayers> {
         RwLockWriteGuard::map(self.write(), |c| &mut c.graphics)
     }
 
     /// What egui outputs each frame.
+    #[inline]
     pub fn output(&self) -> RwLockWriteGuard<'_, Output> {
         RwLockWriteGuard::map(self.write(), |c| &mut c.output)
     }
 
+    #[inline]
     pub(crate) fn frame_state(&self) -> RwLockWriteGuard<'_, FrameState> {
         RwLockWriteGuard::map(self.write(), |c| &mut c.frame_state)
     }
@@ -480,17 +492,19 @@ impl Context {
     ///     // This is fine!
     /// }
     /// ```
-    #[inline(always)]
+    #[inline]
     pub fn input(&self) -> RwLockReadGuard<'_, InputState> {
         RwLockReadGuard::map(self.read(), |c| &c.input)
     }
 
+    #[inline]
     pub fn input_mut(&self) -> RwLockWriteGuard<'_, InputState> {
         RwLockWriteGuard::map(self.write(), |c| &mut c.input)
     }
 
     /// Not valid until first call to [`Context::run()`].
     /// That's because since we don't know the proper `pixels_per_point` until then.
+    #[inline]
     pub fn fonts(&self) -> RwLockReadGuard<'_, Fonts> {
         RwLockReadGuard::map(self.read(), |c| {
             c.fonts
@@ -499,8 +513,20 @@ impl Context {
         })
     }
 
+    #[inline]
     fn fonts_mut(&self) -> RwLockWriteGuard<'_, Option<Fonts>> {
         RwLockWriteGuard::map(self.write(), |c| &mut c.fonts)
+    }
+
+    #[inline]
+    pub fn options(&self) -> RwLockWriteGuard<'_, Options> {
+        RwLockWriteGuard::map(self.write(), |c| &mut c.memory.options)
+    }
+
+    /// Change the options used by the tessellator.
+    #[inline]
+    pub fn tessellation_options(&self) -> RwLockWriteGuard<'_, TessellationOptions> {
+        RwLockWriteGuard::map(self.write(), |c| &mut c.memory.options.tessellation_options)
     }
 }
 
@@ -522,7 +548,7 @@ impl Context {
     pub fn set_fonts(&self, font_definitions: FontDefinitions) {
         if let Some(current_fonts) = &*self.fonts_mut() {
             // NOTE: this comparison is expensive since it checks TTF data for equality
-            if current_fonts.definitions() == &font_definitions {
+            if current_fonts.lock().fonts.definitions() == &font_definitions {
                 return; // no change - save us from reloading font textures
             }
         }
@@ -532,7 +558,7 @@ impl Context {
 
     /// The [`Style`] used by all subsequent windows, panels etc.
     pub fn style(&self) -> Arc<Style> {
-        self.memory().options.style.clone()
+        self.options().style.clone()
     }
 
     /// The [`Style`] used by all new windows, panels etc.
@@ -547,7 +573,7 @@ impl Context {
     /// ctx.set_style(style);
     /// ```
     pub fn set_style(&self, style: impl Into<Arc<Style>>) {
-        self.memory().options.style = style.into();
+        self.options().style = style.into();
     }
 
     /// The [`Visuals`] used by all subsequent windows, panels etc.
@@ -560,7 +586,7 @@ impl Context {
     /// ctx.set_visuals(egui::Visuals::light()); // Switch to light mode
     /// ```
     pub fn set_visuals(&self, visuals: crate::Visuals) {
-        std::sync::Arc::make_mut(&mut self.memory().options.style).visuals = visuals;
+        std::sync::Arc::make_mut(&mut self.options().style).visuals = visuals;
     }
 
     /// The number of physical pixels for each logical point.
@@ -701,25 +727,21 @@ impl Context {
             self.request_repaint();
         }
 
-        self.fonts().end_frame();
-
         {
             let ctx_impl = &mut *self.write();
             ctx_impl
                 .memory
                 .end_frame(&ctx_impl.input, &ctx_impl.frame_state.used_ids);
 
-            let font_image = ctx_impl.fonts.as_ref().unwrap().font_image();
-            let font_image_version = font_image.version;
-
-            if Some(font_image_version) != ctx_impl.latest_font_image_version {
+            let font_image_delta = ctx_impl.fonts.as_ref().unwrap().font_image_delta();
+            if let Some(font_image_delta) = font_image_delta {
                 ctx_impl
                     .tex_manager
                     .0
                     .write()
-                    .set(TextureId::default(), font_image.image.clone().into());
-                ctx_impl.latest_font_image_version = Some(font_image_version);
+                    .set(TextureId::default(), font_image_delta);
             }
+
             ctx_impl
                 .output
                 .textures_delta
@@ -750,14 +772,14 @@ impl Context {
         // shapes are the same, but just comparing the shapes takes about 50% of the time
         // it takes to tessellate them, so it is not a worth optimization.
 
-        let mut tessellation_options = self.memory().options.tessellation_options;
+        let mut tessellation_options = *self.tessellation_options();
         tessellation_options.pixels_per_point = self.pixels_per_point();
         tessellation_options.aa_size = 1.0 / self.pixels_per_point();
         let paint_stats = PaintStats::from_shapes(&shapes);
         let clipped_meshes = tessellator::tessellate_shapes(
             shapes,
             tessellation_options,
-            self.fonts().font_image().size(),
+            self.fonts().font_image_size(),
         );
         self.write().paint_stats = paint_stats.with_clipped_meshes(&clipped_meshes);
         clipped_meshes
@@ -880,12 +902,12 @@ impl Context {
 
     /// Wether or not to debug widget layout on hover.
     pub fn debug_on_hover(&self) -> bool {
-        self.memory().options.style.debug.debug_on_hover
+        self.options().style.debug.debug_on_hover
     }
 
     /// Turn on/off wether or not to debug widget layout on hover.
     pub fn set_debug_on_hover(&self, debug_on_hover: bool) {
-        let mut style = (*self.memory().options.style).clone();
+        let mut style = (*self.options().style).clone();
         style.debug.debug_on_hover = debug_on_hover;
         self.set_style(style);
     }
@@ -956,23 +978,13 @@ impl Context {
                 self.style_ui(ui);
             });
 
-        CollapsingHeader::new("🔠 Fonts")
-            .default_open(false)
-            .show(ui, |ui| {
-                let mut font_definitions = self.fonts().definitions().clone();
-                font_definitions.ui(ui);
-                let font_image = self.fonts().font_image();
-                font_image.ui(ui);
-                self.set_fonts(font_definitions);
-            });
-
         CollapsingHeader::new("✒ Painting")
             .default_open(true)
             .show(ui, |ui| {
-                let mut tessellation_options = self.memory().options.tessellation_options;
+                let mut tessellation_options = self.options().tessellation_options;
                 tessellation_options.ui(ui);
                 ui.vertical_centered(|ui| reset_button(ui, &mut tessellation_options));
-                self.memory().options.tessellation_options = tessellation_options;
+                *self.tessellation_options() = tessellation_options;
             });
     }
 
@@ -1042,6 +1054,13 @@ impl Context {
             .show(ui, |ui| {
                 self.texture_ui(ui);
             });
+
+        CollapsingHeader::new("🔠 Font texture")
+            .default_open(false)
+            .show(ui, |ui| {
+                let font_image_size = self.fonts().font_image_size();
+                crate::introspection::font_texture_ui(ui, font_image_size);
+            });
     }
 
     /// Show stats about the allocated textures.
@@ -1083,8 +1102,12 @@ impl Context {
                                 size *= (max_preview_size.x / size.x).min(1.0);
                                 size *= (max_preview_size.y / size.y).min(1.0);
                                 ui.image(texture_id, size).on_hover_ui(|ui| {
-                                    // show full size on hover
-                                    ui.image(texture_id, Vec2::new(w as f32, h as f32));
+                                    // show larger on hover
+                                    let max_size = 0.5 * ui.ctx().input().screen_rect().size();
+                                    let mut size = Vec2::new(w as f32, h as f32);
+                                    size *= max_size.x / size.x.max(max_size.x);
+                                    size *= max_size.y / size.y.max(max_size.y);
+                                    ui.image(texture_id, size);
                                 });
 
                                 ui.label(format!("{} x {}", w, h));
@@ -1106,8 +1129,8 @@ impl Context {
             *self.memory() = Default::default();
         }
 
-        let num_state = self.memory().data.len();
-        let num_serialized = self.memory().data.count_serialized();
+        let num_state = self.data().len();
+        let num_serialized = self.data().count_serialized();
         ui.label(format!(
             "{} widget states stored (of which {} are serialized).",
             num_state, num_serialized
@@ -1151,13 +1174,10 @@ impl Context {
         ui.horizontal(|ui| {
             ui.label(format!(
                 "{} collapsing headers",
-                self.memory()
-                    .data
-                    .count::<containers::collapsing_header::State>()
+                self.data().count::<containers::collapsing_header::State>()
             ));
             if ui.button("Reset").clicked() {
-                self.memory()
-                    .data
+                self.data()
                     .remove_by_type::<containers::collapsing_header::State>();
             }
         });
@@ -1165,30 +1185,30 @@ impl Context {
         ui.horizontal(|ui| {
             ui.label(format!(
                 "{} menu bars",
-                self.memory().data.count::<menu::BarState>()
+                self.data().count::<menu::BarState>()
             ));
             if ui.button("Reset").clicked() {
-                self.memory().data.remove_by_type::<menu::BarState>();
+                self.data().remove_by_type::<menu::BarState>();
             }
         });
 
         ui.horizontal(|ui| {
             ui.label(format!(
                 "{} scroll areas",
-                self.memory().data.count::<scroll_area::State>()
+                self.data().count::<scroll_area::State>()
             ));
             if ui.button("Reset").clicked() {
-                self.memory().data.remove_by_type::<scroll_area::State>();
+                self.data().remove_by_type::<scroll_area::State>();
             }
         });
 
         ui.horizontal(|ui| {
             ui.label(format!(
                 "{} resize areas",
-                self.memory().data.count::<resize::State>()
+                self.data().count::<resize::State>()
             ));
             if ui.button("Reset").clicked() {
-                self.memory().data.remove_by_type::<resize::State>();
+                self.data().remove_by_type::<resize::State>();
             }
         });