+++ egui_demo_lib/src/easy_mark/easy_mark_highlighter.rs
@@ -5,23 +5,23 @@ use crate::easy_mark::easy_mark_parser;
 /// In practice, the highlighter is fast enough not to need any caching.
 #[derive(Default)]
 pub struct MemoizedEasymarkHighlighter {
-    visuals: egui::Visuals,
+    style: egui::Style,
     code: String,
     output: egui::text::LayoutJob,
 }
 
 impl MemoizedEasymarkHighlighter {
-    pub fn highlight(&mut self, visuals: &egui::Visuals, code: &str) -> egui::text::LayoutJob {
-        if (&self.visuals, self.code.as_str()) != (visuals, code) {
-            self.visuals = visuals.clone();
+    pub fn highlight(&mut self, egui_style: &egui::Style, code: &str) -> egui::text::LayoutJob {
+        if (&self.style, self.code.as_str()) != (egui_style, code) {
+            self.style = egui_style.clone();
             self.code = code.to_owned();
-            self.output = highlight_easymark(visuals, code);
+            self.output = highlight_easymark(egui_style, code);
         }
         self.output.clone()
     }
 }
 
-pub fn highlight_easymark(visuals: &egui::Visuals, mut text: &str) -> egui::text::LayoutJob {
+pub fn highlight_easymark(egui_style: &egui::Style, mut text: &str) -> egui::text::LayoutJob {
     let mut job = egui::text::LayoutJob::default();
     let mut style = easy_mark_parser::Style::default();
     let mut start_of_line = true;
@@ -33,7 +33,7 @@ pub fn highlight_easymark(visuals: &egui::Visuals, mut text: &str) -> egui::text
                 &text[..end],
                 0.0,
                 format_from_style(
-                    visuals,
+                    egui_style,
                     &easy_mark_parser::Style {
                         code: true,
                         ..Default::default()
@@ -50,7 +50,7 @@ pub fn highlight_easymark(visuals: &egui::Visuals, mut text: &str) -> egui::text
             let end = text[1..]
                 .find(&['`', '\n'][..])
                 .map_or_else(|| text.len(), |i| i + 2);
-            job.append(&text[..end], 0.0, format_from_style(visuals, &style));
+            job.append(&text[..end], 0.0, format_from_style(egui_style, &style));
             text = &text[end..];
             style.code = false;
             continue;
@@ -61,7 +61,7 @@ pub fn highlight_easymark(visuals: &egui::Visuals, mut text: &str) -> egui::text
         if text.starts_with('\\') && text.len() >= 2 {
             skip = 2;
         } else if start_of_line && text.starts_with(' ') {
-            // indentation we don't preview indentation, because it is confusing
+            // we don't preview indentation, because it is confusing
             skip = 1;
         } else if start_of_line && text.starts_with("# ") {
             style.heading = true;
@@ -69,15 +69,15 @@ pub fn highlight_easymark(visuals: &egui::Visuals, mut text: &str) -> egui::text
         } else if start_of_line && text.starts_with("> ") {
             style.quoted = true;
             skip = 2;
-            // indentation we don't preview indentation, because it is confusing
+            // we don't preview indentation, because it is confusing
         } else if start_of_line && text.starts_with("- ") {
             skip = 2;
-            // indentation we don't preview indentation, because it is confusing
+            // we don't preview indentation, because it is confusing
         } else if text.starts_with('*') {
             skip = 1;
             if style.strong {
-                // Include the character that i ending ths style:
-                job.append(&text[..skip], 0.0, format_from_style(visuals, &style));
+                // Include the character that is ending this style:
+                job.append(&text[..skip], 0.0, format_from_style(egui_style, &style));
                 text = &text[skip..];
                 skip = 0;
             }
@@ -85,8 +85,8 @@ pub fn highlight_easymark(visuals: &egui::Visuals, mut text: &str) -> egui::text
         } else if text.starts_with('$') {
             skip = 1;
             if style.small {
-                // Include the character that i ending ths style:
-                job.append(&text[..skip], 0.0, format_from_style(visuals, &style));
+                // Include the character that is ending this style:
+                job.append(&text[..skip], 0.0, format_from_style(egui_style, &style));
                 text = &text[skip..];
                 skip = 0;
             }
@@ -94,8 +94,8 @@ pub fn highlight_easymark(visuals: &egui::Visuals, mut text: &str) -> egui::text
         } else if text.starts_with('^') {
             skip = 1;
             if style.raised {
-                // Include the character that i ending ths style:
-                job.append(&text[..skip], 0.0, format_from_style(visuals, &style));
+                // Include the character that is ending this style:
+                job.append(&text[..skip], 0.0, format_from_style(egui_style, &style));
                 text = &text[skip..];
                 skip = 0;
             }
@@ -114,12 +114,16 @@ pub fn highlight_easymark(visuals: &egui::Visuals, mut text: &str) -> egui::text
             .map_or_else(|| text.len(), |i| (skip + i).max(1));
 
         if line_end <= end {
-            job.append(&text[..line_end], 0.0, format_from_style(visuals, &style));
+            job.append(
+                &text[..line_end],
+                0.0,
+                format_from_style(egui_style, &style),
+            );
             text = &text[line_end..];
             start_of_line = true;
             style = Default::default();
         } else {
-            job.append(&text[..end], 0.0, format_from_style(visuals, &style));
+            job.append(&text[..end], 0.0, format_from_style(egui_style, &style));
             text = &text[end..];
             start_of_line = false;
         }
@@ -129,17 +133,17 @@ pub fn highlight_easymark(visuals: &egui::Visuals, mut text: &str) -> egui::text
 }
 
 fn format_from_style(
-    visuals: &egui::Visuals,
+    egui_style: &egui::Style,
     emark_style: &easy_mark_parser::Style,
 ) -> egui::text::TextFormat {
     use egui::{Align, Color32, Stroke, TextStyle};
 
     let color = if emark_style.strong || emark_style.heading {
-        visuals.strong_text_color()
+        egui_style.visuals.strong_text_color()
     } else if emark_style.quoted {
-        visuals.weak_text_color()
+        egui_style.visuals.weak_text_color()
     } else {
-        visuals.text_color()
+        egui_style.visuals.text_color()
     };
 
     let text_style = if emark_style.heading {
@@ -153,7 +157,7 @@ fn format_from_style(
     };
 
     let background = if emark_style.code {
-        visuals.code_bg_color
+        egui_style.visuals.code_bg_color
     } else {
         Color32::TRANSPARENT
     };
@@ -177,7 +181,7 @@ fn format_from_style(
     };
 
     egui::text::TextFormat {
-        style: text_style,
+        font_id: text_style.resolve(egui_style),
         color,
         background,
         italics: emark_style.italics,