+++ egui_glium/src/epi_backend.rs
@@ -47,6 +47,7 @@ pub fn run(app: Box<dyn epi::App>, native_options: &epi::NativeOptions) -> ! {
     let mut painter = crate::Painter::new(&display);
     let mut integration = egui_winit::epi::EpiIntegration::new(
         "egui_glium",
+        painter.max_texture_side(),
         display.gl_window().window(),
         repaint_signal,
         persistence,
@@ -70,8 +71,8 @@ pub fn run(app: Box<dyn epi::App>, native_options: &epi::NativeOptions) -> ! {
                 integration.update(display.gl_window().window());
             let clipped_meshes = integration.egui_ctx.tessellate(shapes);
 
-            for (id, image) in textures_delta.set {
-                painter.set_texture(&display, id, &image);
+            for (id, image_delta) in textures_delta.set {
+                painter.set_texture(&display, id, &image_delta);
             }
 
             // paint: