+++ egui_glium/src/painter.rs
@@ -16,6 +16,7 @@ use {
 };
 
 pub struct Painter {
+    max_texture_side: usize,
     program: glium::Program,
 
     textures: AHashMap<egui::TextureId, Rc<SrgbTexture2d>>,
@@ -27,6 +28,9 @@ pub struct Painter {
 
 impl Painter {
     pub fn new(facade: &dyn glium::backend::Facade) -> Painter {
+        use glium::CapabilitiesSource as _;
+        let max_texture_side = facade.get_capabilities().max_texture_size as _;
+
         let program = program! {
             facade,
             120 => {
@@ -49,11 +53,16 @@ impl Painter {
         .expect("Failed to compile shader");
 
         Painter {
+            max_texture_side,
             program,
             textures: Default::default(),
             #[cfg(feature = "epi")]
             next_native_tex_id: 0,
         }
+    }
+
+    pub fn max_texture_side(&self) -> usize {
+        self.max_texture_side
     }
 
     /// Main entry-point for painting a frame.
@@ -185,9 +194,9 @@ impl Painter {
         &mut self,
         facade: &dyn glium::backend::Facade,
         tex_id: egui::TextureId,
-        image: &egui::ImageData,
+        delta: &egui::epaint::ImageDelta,
     ) {
-        let pixels: Vec<(u8, u8, u8, u8)> = match image {
+        let pixels: Vec<(u8, u8, u8, u8)> = match &delta.image {
             egui::ImageData::Color(image) => {
                 assert_eq!(
                     image.width() * image.height(),
@@ -206,15 +215,29 @@ impl Painter {
         };
         let glium_image = glium::texture::RawImage2d {
             data: std::borrow::Cow::Owned(pixels),
-            width: image.width() as _,
-            height: image.height() as _,
+            width: delta.image.width() as _,
+            height: delta.image.height() as _,
             format: glium::texture::ClientFormat::U8U8U8U8,
         };
         let format = texture::SrgbFormat::U8U8U8U8;
         let mipmaps = texture::MipmapsOption::NoMipmap;
-        let gl_texture = SrgbTexture2d::with_format(facade, glium_image, format, mipmaps).unwrap();
 
-        self.textures.insert(tex_id, gl_texture.into());
+        if let Some(pos) = delta.pos {
+            // update a sub-region
+            if let Some(gl_texture) = self.textures.get(&tex_id) {
+                let rect = glium::Rect {
+                    left: pos[0] as _,
+                    bottom: pos[1] as _,
+                    width: glium_image.width,
+                    height: glium_image.height,
+                };
+                gl_texture.main_level().write(rect, glium_image);
+            }
+        } else {
+            let gl_texture =
+                SrgbTexture2d::with_format(facade, glium_image, format, mipmaps).unwrap();
+            self.textures.insert(tex_id, gl_texture.into());
+        }
     }
 
     pub fn free_texture(&mut self, tex_id: egui::TextureId) {