+++ egui_web/src/glow_wrapping.rs
@@ -40,8 +40,12 @@ impl WrappedGlowPainter {
 }
 
 impl crate::Painter for WrappedGlowPainter {
-    fn set_texture(&mut self, tex_id: egui::TextureId, image: egui::ImageData) {
-        self.painter.set_texture(&self.glow_ctx, tex_id, &image);
+    fn max_texture_side(&self) -> usize {
+        self.painter.max_texture_side()
+    }
+
+    fn set_texture(&mut self, tex_id: egui::TextureId, delta: &egui::epaint::ImageDelta) {
+        self.painter.set_texture(&self.glow_ctx, tex_id, delta);
     }
 
     fn free_texture(&mut self, tex_id: egui::TextureId) {